Improve app component UT
